requirements.txt
streamlit>=1.33.0
openai>=1.0.0 
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
google-auth-oauthlib>=1.2.0
schedule>=1.2.0
python-dateutil>=2.8.0
google-api-python-client>=2.0.0
google-auth-oauthlib==1.2.0